requirements.txt
async-generator==1.10
attrs==21.4.0
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.5.30
cffi==1.15.0
charset-normalizer==2.0.6
coloredlogs==15.0.1
cryptography==37.0.2
deb-pkg-tools==8.4
et-xmlfile==1.1.0
executor==23.2
fasteners==0.16.3
gitdb==4.0.9
GitPython==3.1.26
h11==0.13.0
humanfriendly==10.0
idna==3.2
MouseInfo==0.1.3
openpyxl==3.0.9
outcome==1.1.0
Pillow==9.1.0
pkg_resources==0.0.0
property-manager==3.0
PyAutoGUI==0.9.53
pycparser==2.21
PyGetWindow==0.0.9
pygit==0.1
PyMsgBox==1.0.9
pyOpenSSL==22.0.0
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
PySocks==1.7.1
python3-xlib==0.15
pytweening==1.0.4
PyYAML==6.0
requests==2.26.0
selenium==4.1.5
setproctitle==1.2.2
six==1.16.0
smmap==5.0.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.2.1
tqdm==4.62.3
trio==0.20.0
trio-websocket==0.9.2
typing_extensions==4.2.0
urllib3==1.26.7
verboselogs==1.7
wsproto==1.1.0
